+++ appObjects/FlatCAMGerber.py
@@ -1042,7 +1042,11 @@ class GerberObject(FlatCAMObj, Gerber):
 
         self.marked_rows[:] = []
 
-        aperture = self.ui.apertures_table.item(cw_row, 1).text()
+        try:
+            aperture = self.ui.apertures_table.item(cw_row, 1).text()
+        except AttributeError:
+            self.ui_connect()
+            return
 
         if self.ui.apertures_table.cellWidget(cw_row, 5).isChecked():
             self.marked_rows.append(True)